LPS-58536 [Test] Extract a super class [No logic changes]
